cancellation jobs

+++ database/migrations/2024_03_04_111126_job_posting.php
@@ -21,8 +21,8 @@ return new class extends Migration
             $table->text('job_Remarks')->nullable();
             $table->decimal('job_MinWage', 10, 2)->nullable();
             $table->decimal('job_MaxWage', 10, 2)->nullable();
-            $table->integer('job_Type');
-            $table->integer('job_Edu')->nullable();
+            $table->tinyInteger('job_Type');
+            $table->tinyInteger('job_Edu')->nullable();
             $table->integer('job_Slots')->nullable();
             $table->string('job_Address', 255);
             $table->unsignedBigInteger('barangay_id')->comment('Foreign Key');
